# docs/swagger.yaml
basePath: /v1
definitions:
  models.APIError:
    properties:
      error_message:
        type: string
      error_status:
        type: integer
    type: object
...
  models.Group:
    properties:
      bootdisk:
        type: string
      callbackurl:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      device:
        type: string
      dns:
        type: string
      gateway:
        type: string
      host:
        items:
          $ref: '#/definitions/models.Host'
        type: array
      id:
        type: integer
      image_id:
        type: integer
      ks:
        type: string
      name:
        type: string
      netmask:
        type: string
      ntp:
        type: string
      option:
        items:
          $ref: '#/definitions/models.Option'
        type: array
      options:
        additionalProperties:
          type: string
        type: object
      password:
        type: string
      pool:
        $ref: '#/definitions/models.Pool'
      pool_id:
        type: integer
      syslog:
        type: string
      updated_at:
        type: string
      vlan:
        type: string
    type: object
...
  models.Host:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      domain:
        type: string
      expires_at:
        type: string
      first_seen:
        type: string
      group:
        $ref: '#/definitions/models.Group'
      group_id:
        type: integer
      host_fqdn:
        type: string
      hostname:
        type: string
      id:
        type: integer
      ilo_api_flavour:
        type: string
      ilo_fqdn:
        type: string
      ilo_ip:
        type: string
      ilo_password:
        type: string
      ilo_port:
        type: string
      ilo_user:
        type: string
      ip:
        type: string
      ks:
        type: string
      last_seen:
        type: string
      last_seen_relay:
        description: DHCP parameters
        type: string
      mac:
        type: string
      missing_options:
        type: string
      pool:
        $ref: '#/definitions/models.Pool'
      pool_id:
        type: integer
      progress:
        type: integer
      progresstext:
        type: string
      reimage:
        type: boolean
      updated_at:
        type: string
    type: object
  models.HostForm:
    properties:
      domain:
        type: string
      group_id:
        type: integer
      host_fqdn:
        type: string
      hostname:
        type: string
      ilo_api_flavour:
        type: string
      ilo_fqdn:
        type: string
      ilo_ip:
        type: string
      ilo_password:
        type: string
      ilo_port:
        type: string
      ilo_user:
        type: string
      ip:
        type: string
      ks:
        type: string
      mac:
        type: string
      pool_id:
        type: integer
      progress:
        type: integer
      progresstext:
        type: string
      reimage:
        type: boolean
    type: object
...
  models.Option:
    properties:
      created_at:
        type: string
      data:
        type: string
      deleted_at:
        type: string
      device_class_id:
        type: integer
      host:
        $ref: '#/definitions/models.Host'
      host_id:
        type: integer
      id:
        type: integer
      opcode:
        type: integer
      pool:
        $ref: '#/definitions/models.Pool'
      pool_id:
        type: integer
      priority:
        type: integer
      updated_at:
        type: string
    required:
    - data
    - opcode
    - priority
    type: object
...
  models.Pool:
    properties:
      authorized_vlan:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      end_address:
        type: string
      gateway:
        type: string
      id:
        type: integer
      lease_time:
        type: integer
      managed_reference:
        type: string
      name:
        type: string
      net_address:
        type: string
      netmask:
        type: integer
      only_serve_reimage:
        type: boolean
      start_address:
        type: string
      updated_at:
        type: string
    required:
    - end_address
    - gateway
    - lease_time
    - name
    - netmask
    - start_address
    type: object
...
info:
  contact: {}
  description: VMware Imaging Appliances written in GO with full HTTP-REST API
  title: go-via
  version: "0.1"
paths:
...
  /hosts:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Host'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Get all hosts
      tags:
      - hosts
    post:
      consumes:
      - application/json
      parameters:
      - description: Add a host
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.HostForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Host'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Create a new host
      tags:
      - hosts
  /hosts/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Host ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Remove an existing host
      tags:
      - hosts
    get:
      consumes:
      - application/json
      parameters:
      - description: Host ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Host'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Get an existing Host
      tags:
      - hosts
    patch:
      consumes:
      - application/json
      parameters:
      - description: Host ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update a host
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.HostForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Host'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Update an existing host
      tags:
      - hosts
  /hosts/search:
    post:
      consumes:
      - application/json
      parameters:
      - description: Fields to search for
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.Host'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Host'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Search for an host
      tags:
      - hosts
...
swagger: "2.0"